+++ karen.ts
@@ -4,12 +4,10 @@ import { Ollama } from "npm:ollama";
 import { Command } from "npm:commander";
 import * as Fmt from "jsr:@std/fmt/colors";
 
-import { AssistantService } from "./src/Assistant.ts";
 import { AuthoringService, Edit } from "./src/Authoring.ts";
 import { ConfigStore, KnownConfigParams, remember } from "./src/Config.ts";
 import { Console } from "./src/Console.ts";
-import { EstimateStore } from "./src/Estimate.ts";
-import { Review, ReviewStore } from "./src/Review.ts";
+import { ReviewService, ReviewStore } from "./src/Review.ts";
 import { SettingsV1 } from "./src/Settings.ts";
 import { Store } from "./src/Store.ts";
 
@@ -33,7 +31,7 @@ import {
   touchStorageDir,
 } from "./src/System.ts";
 
-import { ReportingService, ReportStore } from "./src/Reporting.ts";
+// import { ReportingService, ReportStore } from "./src/Reporting.ts";
 
 const console = Console();
 
@@ -209,20 +207,16 @@ export async function main() {
           .catch(() => "none");
         const issues = await Array.fromAsync(IssueStore(storage).keys());
         const projects = await Array.fromAsync(ProjectStore(storage).keys());
-        const boards = await Array.fromAsync(BoardStore(storage).keys());
         const reviews = await Array.fromAsync(ReviewStore(storage).keys());
-        const estimates = await Array.fromAsync(EstimateStore(storage).keys());
         const edits = await Array.fromAsync(EditStore(storage).keys());
         const output = {
           id,
           issues: {
             total: issues.length,
             reviewed: reviews.length,
-            estimated: estimates.length,
             edited: edits.length,
           },
           projects: { total: projects.length },
-          boards: { total: boards.length },
         };
         console.print(output, options.format);
       },
@@ -244,15 +238,6 @@ export async function main() {
 
   list
     .addCommand(
-      new Command("boards")
-        .description("List all stored boards")
-        .option("--details", "Show full board details")
-        .option("-o, --format <format>", "json or yaml", "yaml")
-        .action(ListResource(BoardStore(storage))),
-    );
-
-  list
-    .addCommand(
       new Command("issues")
         .description("List all stored issues")
         .option("--details", "Show full issue details")
@@ -263,7 +248,7 @@ export async function main() {
   list
     .addCommand(
       new Command("my-comments")
-        .description("List all stored comments")
+        .description("List all stored comments published by KAREN")
         .option("--details", "Show full comment details")
         .option("-o, --format <format>", "json or yaml", "yaml")
         .action(ListResource(MyCommentStore(storage))),
@@ -311,23 +296,14 @@ export async function main() {
         }),
     );
 
-  list
-    .addCommand(
-      new Command("estimates")
-        .description("List all stored estimates")
-        .option("--details", "Show full estimate details")
-        .option("-o, --format <format>", "json or yaml", "yaml")
-        .action(ListResource(EstimateStore(storage))),
-    );
-
-  list
-    .addCommand(
-      new Command("reports")
-        .description("List all stored reports")
-        .option("--details", "Show full report details")
-        .option("-o, --format <format>", "json or yaml", "yaml")
-        .action(ListResource(ReportStore(storage))),
-    );
+  // list
+  // .addCommand(
+  //   new Command("reports")
+  //     .description("List all stored reports")
+  //     .option("--details", "Show full report details")
+  //     .option("-o, --format <format>", "json or yaml", "yaml")
+  //     .action(ListResource(ReportStore(storage))),
+  // );
 
   const get = program
     .command("get")
@@ -387,14 +363,6 @@ export async function main() {
   );
 
   get.addCommand(
-    new Command("board")
-      .description("Get stored board")
-      .argument("[board-name]", "JIRA board name")
-      .option("-o, --format <format>", "Output format (json or yaml)", "yaml")
-      .action(GetResource(BoardStore(storage))),
-  );
-
-  get.addCommand(
     new Command("review")
       .description("Get stored review")
       .argument("[issue-key]", "JIRA issue key")
@@ -402,21 +370,13 @@ export async function main() {
       .action(GetResource(ReviewStore(storage))),
   );
 
-  get.addCommand(
-    new Command("estimate")
-      .description("Get stored estimate")
-      .argument("[issue-key]", "JIRA issue key")
-      .option("-o, --format <format>", "Output format (json or yaml)", "yaml")
-      .action(GetResource(EstimateStore(storage))),
-  );
-
-  get.addCommand(
-    new Command("report")
-      .description("Get stored report")
-      .argument("[issue-key]", "JIRA issue key")
-      .option("-o, --format <format>", "Output format (json or yaml)", "yaml")
-      .action(GetResource(ReportStore(storage))),
-  );
+  // get.addCommand(
+  //   new Command("report")
+  //     .description("Get stored report")
+  //     .argument("[project-key]", "JIRA project key")
+  //     .option("-o, --format <format>", "Output format (json or yaml)", "yaml")
+  //     .action(GetResource(ReportStore(storage))),
+  // );
 
   const remove = program
     .command("remove")
@@ -451,15 +411,6 @@ export async function main() {
   );
 
   remove.addCommand(
-    new Command("board")
-      .description("Remove stored board(s)")
-      .argument("[board-name]", "JIRA board name")
-      .option("--all", "Delete all boards")
-      .option("--force", "Delete without confirmation")
-      .action(RemoveResource(BoardStore(storage))),
-  );
-
-  remove.addCommand(
     new Command("review")
       .description("Remove stored review(s)")
       .argument("[issue-key]", "JIRA issue key")
@@ -468,23 +419,14 @@ export async function main() {
       .action(RemoveResource(ReviewStore(storage))),
   );
 
-  remove.addCommand(
-    new Command("estimate")
-      .description("Remove stored estimate(s)")
-      .argument("[issue-key]", "JIRA issue key")
-      .option("--all", "Delete all estimates")
-      .option("--force", "Delete without confirmation")
-      .action(RemoveResource(EstimateStore(storage))),
-  );
-
-  remove.addCommand(
-    new Command("report")
-      .description("Remove stored report(s)")
-      .argument("[issue-key]", "JIRA issue key")
-      .option("--all", "Delete all reports")
-      .option("--force", "Delete without confirmation")
-      .action(RemoveResource(ReportStore(storage))),
-  );
+  // remove.addCommand(
+  //   new Command("report")
+  //     .description("Remove stored report(s)")
+  //     .argument("[project-key]", "JIRA project key")
+  //     .option("--all", "Delete all reports")
+  //     .option("--force", "Delete without confirmation")
+  //     .action(RemoveResource(ReportStore(storage))),
+  // );
 
   const pull = program
     .command("pull")
@@ -573,8 +515,8 @@ export async function main() {
     const reviews = ReviewStore(storage);
     const issueStore = IssueStore(storage);
     const ollama = new Ollama({ ...settings.ollama, model: options.model });
-    const assistant = AssistantService(ollama, storage, settings);
-    await assistant
+    const service = ReviewService(ollama, storage, settings);
+    await service
       .status()
       .catch(console.expect("Failed to connect to Ollama"));
     if (options.all) return await reviewAll(options);
@@ -593,7 +535,7 @@ export async function main() {
 
     async function reviewIssue(issue: Issue, options = settings.printer) {
       console.log("reviewing", issue.key);
-      return await assistant
+      return await service
         .review(issue)
         .then((data) => console.print(data, options.format))
         .catch(console.error);
@@ -621,94 +563,6 @@ export async function main() {
     .option("-o, --format <format>", "Output format (json or yaml)", "yaml")
     .option("--model <model>", "Ollama model to use", settings.ollama.model)
     .action(review);
-
-  async function estimate(
-    key: string,
-    options: {
-      reviewed?: boolean;
-      all?: boolean;
-      force?: boolean;
-      model?: string;
-    } = {},
-  ) {
-    const estimateStore = EstimateStore(storage);
-    const issueStore = IssueStore(storage);
-    const reviewStore = ReviewStore(storage);
-    const ollama = new Ollama(settings.ollama);
-    const assistant = AssistantService(ollama, storage, settings);
-    await assistant
-      .status()
-      .catch(console.expect("Failed to connect to Ollama"));
-
-    if (options.reviewed) return await estimateReviewed(options);
-    if (options.all) return await estimateAll(options);
-
-    const estimate = await estimateStore
-      .get(key)
-      .catch(() => null);
-
-    if (!options.force && estimate) return console.print(estimate);
-
-    return await issueStore
-      .get(key)
-      .catch(console.expect("Issue not found"))
-      .then((issue) => estimateIssue(issue, options));
-
-    async function estimateIssue(
-      issue: Issue,
-      options: { model?: string } = {},
-    ) {
-      console.log("estimating", issue.key);
-      return await assistant
-        .estimate(issue, options)
-        .then((estimate) => estimateStore.put(issue.key, estimate))
-        .then(console.print)
-        .catch(console.error);
-    }
-
-    async function estimateReviewed(
-      options: { force?: boolean; model?: string } = {},
-    ) {
-      const estimatedKeys = await Array
-        .fromAsync(estimateStore.keys())
-        .then((keys) => new Set(keys));
-
-      for await (const key of reviewStore.keys()) {
-        if (!options.force && estimatedKeys.has(key)) continue;
-        await issueStore
-          .get(key)
-          .then((issue) => estimateIssue(issue, options))
-          .catch(console.expect("Issue not found"));
-      }
-
-      return console.info("done");
-    }
-
-    async function estimateAll(
-      options: { force?: boolean; model?: string } = {},
-    ) {
-      const keys = await Array
-        .fromAsync(estimateStore.keys())
-        .then((keys) => new Set(keys));
-      const issues = issueStore.list();
-      for await (const issue of issues) {
-        if (keys.has(issue.key) && !options.force) continue;
-        await estimateIssue(issue, options);
-      }
-      return console.info("done");
-    }
-  }
-
-  program
-    .command("estimate")
-    .description("Estimate story points for an issue using AI")
-    .argument("[issue-key]", "JIRA issue key")
-    .option("--force", "Force a new estimate even if cached")
-    .option("--all", "Estimate all locally stored issues")
-    .option("--reviewed", "Estimate only previously reviewed issues")
-    .option("-o, --format <format>", "Output format (json or yaml)", "yaml")
-    .option("--model <model>", "Ollama model to use")
-    .action(estimate);
 
   async function edit(key: string) {
     const editStore = EditStore(storage);
@@ -834,85 +688,85 @@ export async function main() {
     .option("--created", "Filter by created date")
     .action(prune);
 
-  async function report(
-    key: string,
-    options: {
-      publish?: boolean;
-      force?: boolean;
-      all?: boolean;
-      format?: string;
-      model?: string;
-    } = {},
-  ) {
-    const issueStore = IssueStore(storage);
-    const reportStore = ReportStore(storage);
-    const reporting = ReportingService(storage, settings);
+  // async function report(
+  //   key: string,
+  //   options: {
+  //     publish?: boolean;
+  //     force?: boolean;
+  //     all?: boolean;
+  //     format?: string;
+  //     model?: string;
+  //   } = {},
+  // ) {
+  //   const issueStore = IssueStore(storage);
+  //   const reportStore = ReportStore(storage);
+  //   const reporting = ReportingService(storage, settings);
 
-    if (options.all) return await reportAll(options);
-    if (!key) return console.die("issue key required when not using --all");
+  //   if (options.all) return await reportAll(options);
+  //   if (!key) return console.die("issue key required when not using --all");
 
-    const issue = await issueStore
-      .get(key)
-      .catch(console.expect("Issue not found"));
+  //   const issue = await issueStore
+  //     .get(key)
+  //     .catch(console.expect("Issue not found"));
 
-    return await reportIssue(issue, options);
+  //   return await reportIssue(issue, options);
 
-    async function reportAll(
-      options: {
-        publish?: boolean;
-        force?: boolean;
-        format?: string;
-        model?: string;
-      } = {},
-    ) {
-      const issues = issueStore.list();
-      for await (const issue of issues) await reportIssue(issue, options);
-      return console.info("done");
-    }
+  //   async function reportAll(
+  //     options: {
+  //       publish?: boolean;
+  //       force?: boolean;
+  //       format?: string;
+  //       model?: string;
+  //     } = {},
+  //   ) {
+  //     const issues = issueStore.list();
+  //     for await (const issue of issues) await reportIssue(issue, options);
+  //     return console.info("done");
+  //   }
 
-    async function reportIssue(
-      issue: Issue,
-      options: {
-        publish?: boolean;
-        format?: string;
-        force?: boolean;
-        model?: string;
-        details?: boolean;
-      } = {},
-    ) {
-      console.log("Collecting report for", issue.key);
-      const report = await reporting.collect(issue, options);
-      if (options.publish) {
-        const published = await reporting.publish(report);
-        if (published) {
-          console.info(Fmt.green(`Published ${issue.key}`));
-        } else {
-          console.info(Fmt.yellow(`Already published ${issue.key}`));
-        }
-      }
-      if (options.format === "markdown") {
-        const markdown = await reporting.format(report, { format: "markdown" });
-        return console.log(markdown);
-      }
-      if (options.details) return console.print(report, options.format);
-      return console.print(reportStore.summarize(report), options.format);
-    }
-  }
+  //   async function reportIssue(
+  //     issue: Issue,
+  //     options: {
+  //       publish?: boolean;
+  //       format?: string;
+  //       force?: boolean;
+  //       model?: string;
+  //       details?: boolean;
+  //     } = {},
+  //   ) {
+  //     console.log("Collecting report for", issue.key);
+  //     const report = await reporting.collect(issue, options);
+  //     if (options.publish) {
+  //       const published = await reporting.publish(report);
+  //       if (published) {
+  //         console.info(Fmt.green(`Published ${issue.key}`));
+  //       } else {
+  //         console.info(Fmt.yellow(`Already published ${issue.key}`));
+  //       }
+  //     }
+  //     if (options.format === "markdown") {
+  //       const markdown = await reporting.format(report, { format: "markdown" });
+  //       return console.log(markdown);
+  //     }
+  //     if (options.details) return console.print(report, options.format);
+  //     return console.print(reportStore.summarize(report), options.format);
+  //   }
+  // }
 
-  program
-    .command("report")
-    .description("Generate a report for an issue using AI review and estimate")
-    .argument("[issue-key]", "JIRA issue key")
-    .option("--publish", "Publish report as a comment on the issue")
-    .option("--force", "Force a new review and estimate")
-    .option("--all", "Report on all locally stored issues")
-    .option(
-      "-o, --format <format>",
-      "Output format (json, yaml, markdown)",
-      "yaml",
-    )
-    .option("--model <model>", "Ollama model to use")
-    .action(report);
+  // program
+  //   .command("report")
+  //   .description("Generate a report for an issue using AI review and estimate")
+  //   .argument("[issue-key]", "JIRA issue key")
+  //   .option("--publish", "Publish report as a comment on the issue")
+  //   .option("--force", "Force a new review and estimate")
+  //   .option("--all", "Report on all locally stored issues")
+  //   .option(
+  //     "-o, --format <format>",
+  //     "Output format (json, yaml, markdown)",
+  //     "yaml",
+  //   )
+  //   .option("--model <model>", "Ollama model to use")
+  //   .action(report);
 
   await program.parseAsync();
 }
